+++ virtualizarr/tests/test_manifests/test_group.py
@@ -53,18 +53,18 @@ class TestManifestGroup:
         assert list(manifest_group) == [var]
 
     def test_nbytes_empty(self):
-        assert ManifestGroup().nbytes == 0
+        assert ManifestGroup().nbytes_virtual == 0
 
     def test_nbytes(self, manifest_array):
         a, b = manifest_array(), manifest_array()
         manifest_group = ManifestGroup(arrays={"a": a, "b": b})
-        assert manifest_group.nbytes == a.nbytes_virtual + b.nbytes_virtual
+        assert manifest_group.nbytes_virtual == a.nbytes_virtual + b.nbytes_virtual
 
     def test_nbytes_nested(self, manifest_array):
         a, b = manifest_array(), manifest_array()
         subgroup = ManifestGroup(arrays={"b": b})
         manifest_group = ManifestGroup(arrays={"a": a}, groups={"sub": subgroup})
-        assert manifest_group.nbytes == a.nbytes_virtual + b.nbytes_virtual
+        assert manifest_group.nbytes_virtual == a.nbytes_virtual + b.nbytes_virtual
 
     def test_manifest_repr(self, manifest_array):
         marr = manifest_array(shape=(5, 2), chunks=(5, 2))